reactor: v2: ExceptionPage move to components/
