+++ src/components/Exception/index.js
@@ -1,0 +1,2 @@
+import ExceptionPage from './ExceptionPage.vue'
+export default ExceptionPage
